+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(algebraic.mle)
+
+test_check("algebraic.mle")